refactor: перенос страниц startup и results в page, связка этапов игры
