+++ packages/client/src/pages/GamePage/ui/Game/Game.module.scss
@@ -1,0 +1,10 @@
+.gamePage {
+  width: 90vw;
+  height: 90vh;
+  margin: auto;
+}
+
+.gameField {
+  width: 100%;
+  height: 100%;
+}
