+++ packages/client/src/pages/GamePage/ui/GameResult/GameResults.module.scss
@@ -10,7 +10,7 @@
   text-align: center;
 }
 
-.trophy {
+.icon {
   margin-bottom: 2px;
   font-size: 48px;
 }
