+++ packages/client/src/pages/GamePage/ui/GameStartup/GameStartup.module.scss
