+++ packages/client/src/pages/GamePage/ui/GameStartup/GameStartup.tsx
@@ -1,11 +1,15 @@
 import { Button } from 'antd';
-import { NavLink } from 'react-router-dom';
+import { FC } from 'react';
 import RulesImg from '@/shared/assets/images/rules.png';
-import { getPageUrl } from '@/shared/config';
 import { LogoWithModal } from '@/widgets/LogoWithModal';
-import cls from './GameStartupPage.module.scss';
+import { GamePageStages } from '../../model/gamePageData';
+import cls from './GameStartup.module.scss';
 
-export const GameStartupPage = () => {
+interface Props {
+  changeStage: (stage: GamePageStages) => void;
+}
+
+export const GameStartup: FC<Props> = ({ changeStage }) => {
   return (
     <LogoWithModal
       open
@@ -23,9 +27,12 @@ export const GameStartupPage = () => {
         </div>
 
         <div className={cls.btnWrapper}>
-          <NavLink to={getPageUrl('game')}>
-            <Button type="primary">Start</Button>
-          </NavLink>
+          <Button
+            type="primary"
+            onClick={() => changeStage(GamePageStages.game)}
+          >
+            Start
+          </Button>
         </div>
       </div>
     </LogoWithModal>
