--- packages/client/src/pages/GameResultsPage/index.ts
@@ -1,1 +1,0 @@
-export { GameResultsPage } from './ui/GameResultsPage';
