--- packages/client/src/pages/GameStartupPage/index.ts
@@ -1,1 +1,0 @@
-export { GameStartupPage } from './ui/GameStartupPage';
