+++ packages/client/src/widgets/Layout/ui/Layout.tsx
@@ -36,8 +36,6 @@ export const Layout = () => {
         <NavLink to={getPageUrl('login')}>Login</NavLink>
         <NavLink to={getPageUrl('forum')}>Forum</NavLink>
         <NavLink to={getPageUrl('game')}>Game</NavLink>
-        <NavLink to={getPageUrl('game-results')}>Game results</NavLink>
-        <NavLink to={getPageUrl('game-startup')}>Game startup</NavLink>
         <NavLink to={getPageUrl('leaderboard')}>Leaderboard</NavLink>
         <NavLink to={getPageUrl('not-found')}>Not Found Page</NavLink>
         <NavLink to={getPageUrl('server-error')}>Server Error Page</NavLink>
